Rename class_registry.py to class_registry_functions. Add class_registry.py and class_registry.REGISTRY as registry variable. Initialise class_registry.REGISTRY on startup.
