+++ dionysus_app/class_functions.py
@@ -5,7 +5,9 @@ Functions for creating, editing, dealing with classes.
 import time
 from pathlib import Path
 
-from dionysus_app.class_registry import classlist_exists, register_class
+import class_registry
+
+from dionysus_app.class_registry_functions import classlist_exists, register_class
 from dionysus_app.data_folder import DataFolder, CLASSLIST_DATA_FILE_TYPE
 from dionysus_app.file_functions import convert_to_json
 from dionysus_app.UI_functions import clean_for_filename, input_is_essentially_blank
